Implement CGRA Task Allocation Framework
- Removed the resource-heavy.mlir test file as it is no longer needed.
- Introduced an abstract base class for CGRA task allocation in Allocation.h, defining the interface for mapping Taskflow tasks onto a 2D multi-CGRA grid.
- Added a new test case (allocation-with-resource-binding.mlir) to validate the allocation of CGRAs to tasks with pre-annotated resource-binding attributes, ensuring correct handling of dependencies and task allocation information.
